Move void protection logic to separated class
